refactor: put back the action and argument parsing in the same place
